server/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyjwt==2.6.0
bcrypt==4.0.1
psutil==5.9.5
requests==2.30.0
aiofiles==23.1.0
jinja2==3.1.2
python-dotenv==1.0.0
spacy==3.5.3
torch==2.0.1
numpy==1.24.3
pandas==2.0.1
